page works, still have to add custom error for password which i will do in another story through password-strength library
