+++ libs/shared/angular/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/shared-angular.module';
 export * from './lib/auth.guard';
 export * from './lib/loggedin.guard';
+export * from './lib/passwordValidator.service';
